control_system_dir/requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
distlib==0.3.1
Django==3.0.8
django-debug-toolbar==2.2
django-filter==2.3.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
filelock==3.0.12
graphene==2.1.8
graphene-django==2.11.0
graphql-core==2.3.2
graphql-relay==2.0.1
pipenv==2020.6.2
promise==2.3
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
Unidecode==1.1.1
virtualenv==20.0.25
virtualenv-clone==0.5.4